pyproject.toml
[project]
name = "events-project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.6.2",
    "coverage>=7.13.4",
    "django>=4.2",
    "django-celery-beat>=2.9.0",
    "django-environ>=0.13.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=25.1.0",
    "openpyxl>=3.1.5",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "redis>=7.3.0",
    "requests>=2.32.5",
    "whitenoise[brotli]>=6.12.0",
]
[tool.ruff]
# Длина строки
line-length = 100
# Целевая версия Python
target-version = "py311"

[tool.ruff.lint]
# Список проверок:
# E, W - ошибки и предупреждения pycodestyle
# F - проверки Pyflakes (неиспользуемые импорты и т.д.)
# I - сортировка импортов (замена isort)
# B - проверки flake8-bugbear (поиск потенциальных багов)
# DJ - специфичные правила для Django
select = ["E", "W", "F", "I", "B", "DJ"]
ignore = ["D100", "D101"]

[tool.ruff.lint.isort]
# Группировка импортов: сначала стандартные, потом сторонние, потом Django
section-order = ["future", "standard-library", "third-party", "django", "first-party", "local-folder"]

[tool.ruff.lint.isort.sections]
"django" = ["django", "rest_framework"]

[tool.ruff.format]
# Настройки форматирования (аналог Black)
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "ruff>=0.15.6",
]
